test(profiles): split command suite out of resolution tests
ProfileResolutionTests.swift had grown past the 350-line
ceiling (AGENTS.md §2.1). The Profile commands suite moves to
ProfileCommandTests.swift; monitor-change matching and space
placement stay. Helpers are duplicated per file, matching the
test-suite convention.
